equity_market_monitoring/requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
plotly==5.18.0
sqlalchemy==2.0.25
pymysql==1.1.0
pyyaml==6.0.1
python-dotenv==1.0.0
